+++ lib/moxl/src/Moxl/Xec/Action/Presence/Unavailable.php
@@ -20,6 +20,7 @@ class Unavailable extends Action
 
     public function handle($stanza, $parent = false)
     {
+        $this->pack($this->_to);
         $this->deliver();
     }